The type of id field was changed

+++ link-analysis-model/src/main/java/org/linkAnalysis/model/dao/Dao.java
@@ -38,5 +38,5 @@ public interface Dao<T extends DomainEntity> {
      * @param id
      * @return
      */
-    T get(Long id);
+    T get(int id);
 }
